Rename authentication to auth context
