+++ App.tsx
@@ -4,7 +4,7 @@ import * as SplashScreen from "expo-splash-screen";
 
 import { BaseNavigator } from "@navigators/BaseNavigator";
 import { ThemeProvider } from "@contexts/theme";
-import { AuthProvider } from "@contexts/authentication";
+import { AuthProvider } from "@contexts/auth";
 
 import "expo-dev-client";
